% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{extract_gofchr}
\alias{extract_gofchr}
\title{Extracting GOFs from Models for Footnote using \code{\link[texreg]{extract}()} function}
\usage{
extract_gofchr(m, m.names = NULL, gof.extracts = c("AIC", "Adj. R$^2$",
  "Log Likelihood"), ...)
}
\arguments{
\item{m}{Single model object or the list of model objects.}

\item{m.names}{The names for models (character vector). The default is \code{NULL}, but required when there are multiple models in the input.}

\item{gof.extracts}{The GOF measures to be included in the output (character/character vector). The default is \code{c("AIC", "Adj. R$^2$", "Log Likelihood")}. If set \code{"all"}, then all GOFs extracted by \code{\link[texreg]{extract}()} function are included in texts.}

\item{...}{Additional arguments passed to \code{\link[texreg]{extract}()} function.}
}
\value{
A \code{character} vector, including GOFs and number of observations of results.
}
\description{
Extracting GOFs from Models for footnote using \code{\link[texreg]{extract}()} function from \code{texreg} package.
}
