% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_pred.R
\name{ilogit}
\alias{ilogit}
\title{Inverse Logit Transformation}
\usage{
ilogit(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\code{exp(x)/(1+exp(x))}
}
\description{
Computes the inverse logit transformation. 
The copy of \code{\link[faraway]{ilogit}()} function.
}
