% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_est.R
\name{matrix_coefci}
\alias{matrix_coefci}
\title{Export Matrix (Data.Frame) of Coefficients and Confidence Interval}
\usage{
matrix_coefci(m, level = 0.95, vcov. = NULL, sims = NULL)
}
\arguments{
\item{m}{A single model object.}

\item{level}{Confidence interval pecent (0-1).}

\item{vcov.}{Alternative variance covariance matrix.}

\item{sims}{Number of iterations for simulation (integer). If \code{NULL}, the coefficients and confidence intervals are extracted analytically.}
}
\description{
Export Matrix (Data.Frame) of Coefficients and Confidence Interval
}
