% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_footnote.R
\name{plot_footnote}
\alias{plot_footnote}
\title{Adding footnote to the \code{ggplot} object}
\usage{
plot_footnote(p, note, fontsize = NULL, fontcol = "black",
  align = "right", distance.from.side = 0.05, distance.from.bottom = 0.75,
  bottom.expand.rate = NULL, show.plot = TRUE)
}
\arguments{
\item{p}{\code{ggplot} object.}

\item{note}{The content of footnote (character). Line break is allowed by using \code{\\n}.}

\item{fontsize}{The size of font. If \code{NULL} (default), the size is set to the the font size in \code{text} setting of \code{p} - 1.}

\item{fontcol}{The color of the font. The default is \code{"black"}.}

\item{align}{The alignment of the footnote text. Either \code{"right"} or \code{"left"}.}

\item{distance.from.side}{The horizontal distance of notes from the edge of graph space by the proportion of graph width (numeric: 0-1). The default is \code{0.05}. The distance is measured from the side specified in \code{align}.}

\item{distance.from.bottom}{The vertical distance of notes from the bottom of graph bottom by the proportion of bottom graph margin height (numeric: 0-1). The default is \code{0.75}.}

\item{bottom.expand.rate}{The expansion rate of the bottom margin of the graph to incorporate footnote (numeric). The value of \code{1} indicates no expansion. If \code{NULL} (default), it is set to the number of lines in \code{note} + 1.}

\item{show.plot}{Print the plot at the end of function (boulean). The default is \code{TRUE}.}
}
\value{
\code{gtable} object. It is impossible to add aditional ggplot elements to this object. Plot can be viewed by using either \code{\link[grid]{grid.draw}()} or \code{\link[graphics]{plot}()} function.
}
\description{
This function adds footnote to the \code{ggplot} object and returns \code{gtable} object.
}
\examples{
## Load Data
library(pscl)
data(vote92)

## Recode Variables
vote92$voteBush <- as.numeric(
  factor(vote92$vote,levels=c("Clinton","Bush")))*1 - 1
vote92$bushdis <- sqrt(vote92$bushdis)
vote92$clintondis <- sqrt(vote92$clintondis)

## Estimate Logistic Regression
fm <- formula(voteBush ~ dem + rep +
                clintondis + bushdis +
                persfinance + natlecon)
m <- glm(fm, data = vote92,
         family = binomial("logit"))

## Basic Graph
p <- plot_coef(m)

## Add Footnote
p_wfoot <- plot_footnote(p,"Votes for Ross Perot are eliminated from analysis.")
p_wfoot

}
