% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{plot_save}
\alias{plot_save}
\alias{pdf_save}
\alias{bmp_save}
\alias{png_save}
\alias{jpeg_save}
\alias{tiff_save}
\title{Save Plots}
\usage{
plot_save(p, w = NULL, h = NULL, res = 120, file = NULL, dir = NULL,
  format = "pdf", show.plot = TRUE, ...)

pdf_save(p, w = 8, h = 6, file = NULL, dir = NULL, ...)

bmp_save(p, w = 800, h = 600, res = 120, file = NULL, dir = NULL, ...)

png_save(p, w = 800, h = 600, res = 120, file = NULL, dir = NULL, ...)

jpeg_save(p, w = 800, h = 600, res = 120, file = NULL, dir = NULL,
  ...)

tiff_save(p, w = 800, h = 600, res = 120, file = NULL, dir = NULL,
  ...)
}
\arguments{
\item{p}{The \code{ggplot} object.}

\item{w}{The width of the plot ouput. If \code{NULL} (default), \code{8} for \code{format = "pdf"}, \code{800} for \code{format \%in\% c("bmp", "png", "jpeg", "tiff")}.}

\item{h}{The height of the plot output. If \code{NULL} (default), \code{6} for \code{format = "pdf"}, \code{600} for \code{format \%in\% c("bmp", "png", "jpeg", "tiff")}.}

\item{res}{The resolution of the plot ouput. Used if \code{format \%in\% c("bmp", "png", "jpeg", "tiff")}. The default value is \code{120}.}

\item{file}{The (path and) name of the file (including extention). If \code{NULL} (default), the file is saved to the current working directory with the name of the object \code{p}.}

\item{dir}{The directory to save the plot. If \code{NULL} (default), the file is saved to the current working directory (or the path included in \code{file} argument). If both \code{dir} and \code{file} are specified, then the file is saved to the combined path of \code{dir} and \code{file}.}

\item{format}{The format of the saving file. It can be chose from \code{"pdf", "bmp", "png", "jpeg", "tiff")}. Default value is \code{"pdf"}.}

\item{show.plot}{Print the plot at the end of function (boulean). The default is \code{TRUE}.}

\item{...}{Additional arguments passed to the graphic device function specified in \code{format}.}
}
\description{
Save and export plots with the specified names. Use \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{bmp}}, \code{\link[grDevices]{png}}, \code{\link[grDevices]{jpeg}} or \code{\link[grDevices]{tiff}} to save the plot ouput to the specified directory.
}
\examples{
## Load Data
library(pscl)
data(vote92)

## Recode Variables
vote92$voteBush <- as.numeric(
  factor(vote92$vote,levels=c("Clinton","Bush")))*1 - 1
vote92$bushdis <- sqrt(vote92$bushdis)
vote92$clintondis <- sqrt(vote92$clintondis)

## Estimate Logistic Regression
fm <- formula(voteBush ~ dem + rep +
                clintondis + bushdis +
                persfinance + natlecon)
m <- glm(fm, data = vote92,
         family = binomial("logit"))

## Basic Graph
p <- plot_coef(m)
p

## Save Graph
plot_save(p)

}
