% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_interact.R
\name{simu_interact}
\alias{simu_interact}
\title{Simulate conditional effect from models with interaction terms}
\usage{
simu_interact(m, var1, var1.label = NULL, moveprof = NULL,
  var2 = NULL, steps = 100, y.label = "auto", show.ci = TRUE,
  level.ci = 0.95, vcov.est = NULL, robust.type = "HC1",
  cluster.var = NULL, iterate.num = 1000, iterate.seed = 578,
  rawbeta = NULL, ...)
}
\arguments{
\item{m}{Single model object.}

\item{var1}{Variable of coeffcients to be simulated.}

\item{var1.label}{Text label for \code{var1} (optional).}

\item{moveprof}{Data.frame object that specifies moving values for variables included in interaction terms other than \code{var1}.
The first variable in data.frame is considered as the primary variable for simulation. 
All variables not included in \code{moveprof} are assumed not to be interacted.
Alternatively, you can specify \code{var2} if only two variables are interacted in the model and 
you want to simulate coefficients for full range of values for the second variable. (Note: You must give one of \code{moveprof} and \code{var2}).}

\item{var2}{Character for the second variable name in interaction term. Must be given is \code{moveprof} is \code{NULL}.}

\item{steps}{Maximum number of values to be simulated. (Applied if \code{var2} is not \code{NULL} and \code{moveprof} is \code{NULL}).}

\item{y.label}{The label of the dependent variable (optional, character).}

\item{show.ci}{Show confidence interval (boulean). The default is \code{TRUE}.}

\item{level.ci}{The level used for confidence interval (numeric: 0-1). The default is \code{0.95}.}

\item{vcov.est}{Variance-covariance matrix to draw coefficents. 
If \code{"robust"}, use robust variance-covariance matrix (also see \code{robust.type}).
If \code{"cluster"}, use cluster robust variance-covariance matrix (also see \code{cluster.var}).
If \code{"boot"}, draw beta by non-parametric bootstrap method using \code{\link[car]{Boot}} function.
If \code{"rawbeta"}, use pre-simulated beta given in \code{rawbeta} argument.
If \code{NULL} (default), use the standard variance covariance matrix stored the model. You can also directly set the variance-covariance matrix created by \code{\link[stats]{vcov}()} or \code{\link[sandwich]{vcovHC}()}.}

\item{robust.type}{The type of leverage adjustment passed to \code{[sandwich]{vcovHC}} (applied only when \code{vcov.est=="robust"}).}

\item{cluster.var}{A \code{vector}, \code{matrix}, or \code{data.frame} of cluster variables, where each column is a separate variable. Alternatively, a formula specifying the cluster variables to be used (see Details in \code{[multiwayvcov]{cluster}}. Applied only when \code{vcov.est=="cluster"}.)}

\item{iterate.num}{The number of iteration in simulation.}

\item{iterate.seed}{The seed value for random number generator used for the draws from multivariate normal distribution.}

\item{rawbeta}{The matrix of pre-simulated beta. Columns are variables, raws are simulated cases. Used only when \code{vcov.est=="rawbeta"}.}

\item{...}{Additional arguments passed to \code{[sandwich]{vcovHC}}, \code{[multiwayvcov]{cluster}}, or \code{\link[car]{Boot}} depending on the value of \code{vcov.est}.}
}
\value{
A list of: 
\itemize{
 \item{\code{predsum}}{Summary Predictions Table}
 \item{\code{profile}}{Profile Used for Predictions}
 \item{\code{predres}}{Raw Predictions}
 \item{\code{formula}}{Estimation Formula}
 \item{\code{y.label}}{Dependent Variable Label}
 \item{\code{family}}{Estimation Method Family}
 \item{\code{type}}{Output Type}
 \item{\code{moveprof}}{Profile of Moving Values}
 \item{\code{var1.label}}{The Label of Target variable}
 \item{\code{var2}}{Primary condition variable}
}
}
\description{
Compute interaction effects through Monte-Carlo Simulation. Wrapper function of \code{\link{simu_pred}}.
}
\examples{

## Load Data
library(pscl)
data(vote92)

## Recode Variables
vote92$voteBush <- as.numeric(
factor(vote92$vote,levels=c("Clinton","Bush")))*1 - 1
vote92$bushdis <- sqrt(vote92$bushdis)
vote92$clintondis <- sqrt(vote92$clintondis)

## Estimate Logistic Regression with Interaction
fm <- formula(voteBush ~ dem*clintondis + 
                rep + bushdis +
                persfinance + natlecon)
m <- glm(fm, data = vote92,
         family = binomial("logit"))
         
# Moving Values
moveprof <- data.frame(clintondis = seq(0,4,length=50))
# Simulation
interactprof <- simu_interact(m, "dem", moveprof=moveprof,
                              var1.label = "Being Democrat",
                              y.label = "Bush Vote")
# Plot
plot_interact(interactprof, 
              label.var2="Ideological Distance from Clinton")

# Alternative Way
interactprof <- simu_interact(m, "dem", var2 = "clintondis",
                              var1.label = "Being Democrat",
                              y.label = "Bush Vote")
plot_interact(interactprof,
              label.var2="Ideological Distance from Clinton")

}
