% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_pred.R
\name{simu_pred}
\alias{simu_pred}
\title{Exporting Simulated Predictions}
\usage{
simu_pred(m, profile = NULL, type = "response", at = "means",
  show.ci = TRUE, level.ci = 0.95, est.vcov = "robust",
  robust.type = "HC1", iterate.num = 1000, iterate.seed = 578)
}
\arguments{
\item{m}{Single model object or the list of model objects.}

\item{profile}{The case profile  (\code{matrix} or \code{data.frame}). Column names are variables in the model and prediction is made for each row. Any missing variable in the profile is substituted by the rule defined by \code{at}. The default is \code{NULL} (see \code{at}).}

\item{type}{If \code{response} (default), the output is transformed to the response format}

\item{show.ci}{Show confidence interval (boulean). The default is \code{TRUE}.}

\item{level.ci}{The level used for confidence interval (numeric: 0-1). The default is \code{0.95}.}

\item{est.vcov}{Variance-covariance matrix. The default is \code{"robust"} (see \code{robust.type}). If \code{NULL}, use the standard variance covariance matrix stored the model. You can also directly set the variance-covariance matrix created by \code{\link[stats]{vcov}()} or \code{\link[sandwich]{vcovHC}()}.}

\item{robust.type}{The type of robust standard error (applied only when \code{est.vcov=="robust"}).}

\item{iterate.num}{The number of iteration in simulation.}

\item{iterate.seed}{The seed value for random number generator.}
}
\value{
A list of: 
\itemize{
 \item{\code{predsum}}{Summary Predictions}
 \item{\code{profile}}{Profile Used for Predictions}
 \item{\code{predres}}{Raw Predictions}
}
}
\description{
Generate prediction through Monte Carlo simulation.
}
