% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_interact.R
\name{summary.simuinteract}
\alias{summary.simuinteract}
\title{Summary Class for \code{simuinteract} Object}
\usage{
\method{summary}{simuinteract}(object, ..., digits = 3, print.rows = 15)
}
\arguments{
\item{object}{\code{simuinteract} object}

\item{...}{Additional arguments.}

\item{digits}{Number of digits to show when summarizing data.}

\item{print.rows}{Max number of profile rows to be printed.}
}
\value{
Summary output
}
\description{
Summarizing \code{simuinteract} object
}
