% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_pred.R
\name{summary.simupred}
\alias{summary.simupred}
\title{Summary Class for \code{simupred} Object}
\usage{
\method{summary}{simupred}(object, ..., digits = 3, print.rows = 15)
}
\arguments{
\item{object}{\code{simupred} object}

\item{...}{Additional arguments.}

\item{digits}{Number of digits to show when summarizing data.}

\item{print.rows}{Max number of profile rows to be printed.}
}
\value{
Summary output
}
\description{
Summarizing \code{simupred} object
}
