% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_coef.R
\name{table_coef}
\alias{table_coef}
\title{Plotting Coefficients}
\usage{
table_coef(m, format = "tex", file.name = NULL, dir = NULL,
  m.names = NULL, vcov.est = NULL, digits = 3, drop.intercept = FALSE,
  drop.intercept.names = "(Intercept)", drop.variable.names = NULL,
  caption = NULL, custom.variable.names = NULL, footnote.psymbol = TRUE,
  custom.footnote = NULL, show.table = TRUE, booktabs = TRUE,
  dcolumn = FALSE, use.packages = FALSE, stars = c(0.001, 0.01, 0.05,
  0.1), symbol = "dagger", caption.above = TRUE, single.row = FALSE,
  float.pos = "h!!", fontsize = "footnotesize", include.deviance = FALSE,
  include.smooth = FALSE, ...)
}
\arguments{
\item{m}{Single or the list of model object.}

\item{format}{The format for the table output (characater). The default is \code{"tex"} (LaTex table). Other options are \code{"html"} (HTML table), \code{"doc"} (HTML table saved as Microsof Word file), and \code{"screen"} (The R Output. Cannot be saved.)}

\item{file.name}{The name of the file to be saved (character). The default is \code{NULL} (The output not saved to file). Extension is automatically added, so do not include it.}

\item{dir}{The directory of the file to be saved. Only applicable when file.name is not \code{NULL}.}

\item{m.names}{The set of names that identifies each element in \code{m}. Considered if \code{m} is a list of models. The length of the vector must correspond with the length of \code{m}. If \code{NULL} (default), each element \code{i} is temporarily named as \code{Model i}.}

\item{vcov.est}{Single or the list of the alternative variance-covariance matrix to be used for the model object \code{m}. Ignored if \code{NULL} (default) or the \code{m} is not model object. Must have the same length as \code{m} if it is a list.}

\item{digits}{Rounding decimal points for values in the table.}

\item{drop.intercept}{Drop the intercept from the plot (boulean). If \code{FALSE} (default), intercept included in the plot.}

\item{drop.intercept.names}{The name(s) of intercept (character/character vector). Needed if \code{drop.intercept} is \code{TRUE}. This is used to identify and eliminate intercept variables from the output. Default value is \code{'(Intercept)'}.}

\item{drop.variable.names}{The name(s) of additional variables to drop (character/character vector) from the ouput. The default is \code{NULL}.}

\item{caption}{Table title caption (character). The default is to include no title caption.}

\item{custom.variable.names}{List of alternative variable names in the output (character vector). The default is \code{NULL}. This is applied AFTER \code{drop.intercept} and \code{drop.variable.names} are applied, thus you don't need the names for dropped variables.}

\item{custom.footnote}{Custom footnote (character). The default is \code{NULL}. If assigned, footnote are added to the plot by \code{\link{plot_footnote}} function, and the function exports \code{gtable} object. Note that \code{gtable} object is less customizable than \code{ggplot} object. If it is also the case that \code{footnote.gof == TRUE}, custom footnote will be added as the new line after the GOF footnote.}

\item{show.table}{Print the table at the end of function (boulean). The default is \code{TRUE}.}

\item{...}{Other options in \code{texreg} function.}
}
\value{
Output from \code{texreg} function.
}
\description{
Drawing the coefficient table from model object(s). The wrapper function of \code{texreg}.
}
