% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{viseffect.theme}
\alias{viseffect.theme}
\title{Default \code{ggplot} theme for \code{viseffect} package.}
\usage{
viseffect.theme()
}
\description{
Default \code{ggplot} theme for \code{viseffect} package.
}
