% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plotting Coefficients}
\usage{
plot_coef(
  m,
  m.names = NULL,
  order.variable = "original",
  odds = FALSE,
  vcov.est = NULL,
  robust.type = "HC1",
  cluster.var = NULL,
  boot.sims = 300,
  boot.seed = 578,
  boot.ncores = 1,
  show.ci = TRUE,
  level.ci = 0.95,
  drop.intercept = FALSE,
  drop.intercept.names = "(Intercept)",
  drop.variable.names = NULL,
  point.shape = 16,
  point.size = 1.5,
  point.color = "black",
  ci.linetype = 1,
  ci.size = 0.5,
  ci.height = 0.2,
  overlap.names = "m.names",
  overlap.gapwidth = 0.5,
  overlap.shape.index = NULL,
  overlap.color.index = NULL,
  overlap.linetype.index = NULL,
  overlap.legend.position = "bottom",
  category.names = NULL,
  category.names.location = "left",
  category.names.angle = 0,
  facet.names = NULL,
  facet.x.scale = "fixed",
  title = NULL,
  y.title = NULL,
  custom.variable.names = NULL,
  custom.x.title = NULL,
  custom.x.lim = NULL,
  custom.x.breaks = NULL,
  custom.themes = NULL,
  footnote.gof = FALSE,
  footnote.gof.nobs = "right",
  footnote.gof.extracts = c("AIC", "Adj. R$^2$", "Log Likelihood"),
  footnote.gof.reorder = NULL,
  footnote.gof.linebreak = TRUE,
  custom.footnote = NULL,
  footnote.caption = FALSE,
  footnote.fontsize = NULL,
  footnote.fontcol = "black",
  footnote.align = "right",
  footnote.distance.from.side = 0.05,
  footnote.distance.from.bottom = 0.75,
  footnote.bottom.expand.rate = NULL,
  show.plot = TRUE,
  flip.plot = FALSE,
  ...
)
}
\arguments{
\item{m}{Single or the list of model object or the \code{matrix} or \code{data.frame} object. If it's a list all elements of the list should be the same class. If model object, \code{\link[stats]{coef}} and \code{\link[lmtest]{coefci}} must be applicable to the object. If  \code{matrix} or \code{data.frame}, then it should include following three columns and row names. Each row should correspond to each variable.
\itemize{
 \item{Row Names: }{Name of Variables}
 \item{First Column: }{Coefficient}
 \item{Second Column: }{Lower limit of confidence interval}
 \item{Third Column: }{Upper limit of confidence interval}
}}

\item{m.names}{The set of names that identifies each element in \code{m}. 
Considered if \code{m} is a list of models. 
The length of the vector must correspond with the length of \code{m}. 
If \code{NULL} (default), each element \code{i} is temporarily named as \code{Model i}.}

\item{order.variable}{Order of coefficients in the plot(character/numeric vector). 
\code{"original"} (default) preserves the original order of the variables. 
\code{"cfdescend"} plots by the descending order of coefficient values. 
\code{"cfascend"} plots by the ascending order of coefficient values.
Alternatively, you can also specify the order of variables by numeric vector 
(applied after \code{drop.intercept} and \code{drop.variable.names} are applied, 
thus you don't need the names for dropped variables).}

\item{odds}{Use odds ratio instead of coefficient in the output (boulean). 
The default is \code{FALSE}. 
If \code{TRUE}, the exponent of the coefficients will be plotted.}

\item{vcov.est}{Single or a list of alternative variance-covariance matrix. 
Each element must be one of raw variance-covariance matrix, \code{"robust"}, 
\code{"cluster"}, \code{"boot"}, or \code{NULL} (default).
If \code{"robust"}, robust standard error (also see \code{robust.type}).
If \code{"cluster"}, cluster robust standard error (also see \code{cluster.var}).
if \code{"boot"}, bootstrapped standard error calculated by \code{\link[car]{Boot}} 
function is used.
Ignored if \code{NULL} (default) or the \code{m} is not model object. 
Must have the same length as \code{m} if it is a list.}

\item{robust.type}{The type of leverage adjustment passed to \code{\link[sandwich]{vcovHC}} 
(applied only when \code{vcov.est=="robust"}).}

\item{cluster.var}{Single or a list of \code{vector}, \code{matrix}, or \code{data.frame} 
of cluster variables, where each column is a separate variable. Alternatively, 
a formula specifying the cluster variables to be used 
(see details in \code{\link[multiwayvcov]{cluster.vcov}}. Applied only when \code{vcov.est=="cluster"}.)}

\item{boot.sims}{Number of iterations if bootstrap is used.}

\item{boot.seed}{Random number seed if bootstrap method is used.}

\item{boot.ncores}{Number of cores to parallelize bootstrap. The default is \code{1}. Use \code{"auto"} to automatically detect number of cores in computer.}

\item{show.ci}{Show confidence interval (boulean). The default is \code{TRUE}.}

\item{level.ci}{The level used for confidence interval (numeric: 0-1). 
The default is \code{0.95}.}

\item{drop.intercept}{Drop the intercept from the plot (boulean). 
If \code{FALSE} (default), intercept included in the plot.}

\item{drop.intercept.names}{The name(s) of intercept (character/character vector). 
Needed if \code{drop.intercept} is \code{TRUE}. 
This is used to identify and eliminate intercept variables from the output. 
Default value is \code{"(Intercept)"}.}

\item{drop.variable.names}{The name(s) of additional variables to drop 
(character/character vector) from the ouput. The default is \code{NULL}.}

\item{point.shape}{Shape of the point outputs (numeric/character). 
The default is \code{16} (filled circle).}

\item{point.size}{Size of point outputs (numeric). The default is \code{1.5}.}

\item{point.color}{Color of point (and ci) outputs (numeric). The default is \code{"black"}.}

\item{ci.linetype}{The line type of confidence interval outputs (numeric). 
The default is \code{1}.}

\item{ci.size}{The line width of confidence interval outputs (numeric). 
The default is \code{0.5}.}

\item{ci.height}{The height of the vertical line added to the edge of confidence 
interval outputs. The default is \code{0.2}.}

\item{overlap.names}{Model labels that controls overlapping 
(applied only when there are two or more \code{m}s). 
Default is \code{"m.names"} (use values specified in \code{m.names}). 
You can alternatively give a character vector of the same length as \code{m}.}

\item{overlap.gapwidth}{The gap between overlapped ouputs (numeric). 
The default value is \code{0.5}.}

\item{overlap.shape.index}{The index of shapes for overlapped point ouputs. 
Must be in the same length as the list \code{m}. 
The first element of the vector is the shape for \code{m}, 
then from the second element, the order must correspond with the order in the list \code{m}. 
If \code{NULL}, \code{point.shape} is applied to all classes.}

\item{overlap.color.index}{The index of colors for overlapped point ouputs. 
Must be in the same length as the list \code{m}. 
The first element of the vector is the shape for \code{m}, 
then from the second element, the order must correspond with the order in the list \code{m}. 
If \code{NULL}, \code{point.shape} is applied to all classes.}

\item{overlap.linetype.index}{The index of line types for overlapped confidence interval 
ouputs. Must be in the same length as the list \code{m}. 
The first element of the vector is the shape for \code{m}, then from the second element, 
the order must correspond with the order in the list \code{m}. If \code{NULL}, 
the number corresponding with the order is assigned to each class.}

\item{overlap.legend.position}{The position of the legend for overlapping classess. 
See \code{legend.position} in ggplot theme for possible values. 
The default is \code{"bottom"}.}

\item{category.names}{The categories of variables (factor). 
If not \code{NULL}, the output provides the separate panels for variables in each category. 
The length of the vector must much with the number of variables in \code{m} 
(applied after \code{drop.intercept} and \code{drop.variable.names} are applied, 
thus you don't need the names for dropped variables).}

\item{category.names.location}{The location of category names. Either \code{"left"} or \code{"right"}. The default is \code{"left"}.}

\item{category.names.angle}{The angle of category names (numeric). The default is \code{0} (horizontal).}

\item{facet.names}{If not \code{NULL}, facet models by this identifier. Use \code{"m.names"} to facet each model separately (\code{overlap.names} is forced to \code{NULL}). 
Alternatively, assign character vector of the same length as \code{m} (number of models).}

\item{facet.x.scale}{If \code{"fixed"} (default), all facetted models have same x axis limits. If \code{"free"} the scale is adjusted by each facetted panel.}

\item{title}{Plot title (character). The default is to include no title.}

\item{y.title}{Y axis title (character). The default is to include no axis title.}

\item{custom.variable.names}{Set of alternative variable names in the output (character vector). The default is \code{NULL}. This is applied AFTER \code{drop.intercept} and \code{drop.variable.names} are applied, thus you don't need the names for dropped variables.}

\item{custom.x.title}{Custom name for the X axis title (character). The default is \code{NULL}.}

\item{custom.x.lim}{Custom limit for the X axis. The default is \code{NULL}.}

\item{custom.x.breaks}{Custom breaks for the X axis. The default is \code{NULL}.}

\item{custom.themes}{ggplot themes that overrides the default theme. The default is \code{NULL}.}

\item{footnote.gof}{Include GOF measures in the footnote (boulean). The default is \code{FALSE}. If \code{TRUE}, footnote with GOF measures are added to the plot by \code{\link{extract_gofchr}} and \code{\link{plot_footnote}} function, and the function exports \code{gtable} object. Note that \code{gtable} object is less customizable than \code{ggplot} object.}

\item{footnote.gof.nobs}{The position of "number of observation (N)" to be printed. The default is \code{"right"} (the end of the sentence). \code{"left"} (the start of the sentence) and \code{"none"} (do not print number of observation) are also available.}

\item{footnote.gof.extracts}{GOF measures to be inluded if \code{footnote.gof == TRUE}. See \code{\link{extract_gofchr}} documentation for more details.}

\item{footnote.gof.reorder}{The numeric vector specifying the alternative order of GOF  (e.g., if there are three GOFs, and you want to flip the order of first and second GOF, then input \code{c(2,1,3)}). The length of the vector must correspond with the number of exported GOFs (it may not be the same as the length of \code{footnote.gof.extracts}). If \code{NULL} (default), the original order is used. The order does not include the number of observation (if exported).}

\item{footnote.gof.linebreak}{Include linebreak between models (boulean). If \code{TRUE} (default), linebreak will be included between models. If \code{FALSE}, the ouput will be single line. Only applicable when there are more than one model.}

\item{custom.footnote}{Custom footnote (character). The default is \code{NULL}. If assigned, footnote are added to the plot by \code{\link{plot_footnote}} function, and the function exports \code{gtable} object. Note that \code{gtable} object is less customizable than \code{ggplot} object. If it is also the case that \code{footnote.gof == TRUE}, custom footnote will be added as the new line after the GOF footnote.}

\item{footnote.caption}{Use caption option in ggplot2 to add footnote. If \code{TRUE}, all footnote options below are ignored.}

\item{footnote.fontsize}{The size of font. If \code{NULL} (default), the size is set to the the font size in \code{text} setting of ggplot theme - 1.}

\item{footnote.fontcol}{The color of the font. The default is \code{"black"}.}

\item{footnote.align}{The alignment of the footnote text. Either \code{"right"} or \code{"left"}.}

\item{footnote.distance.from.side}{The horizontal distance of notes from the edge of graph space by the proportion of graph width (numeric: 0-1). The default is \code{0.05}. The distance is measured from the side specified in \code{align}.}

\item{footnote.distance.from.bottom}{The vertical distance of notes from the bottom of graph bottom by the proportion of bottom graph margin height (numeric: 0-1). The default is \code{0.75}.}

\item{footnote.bottom.expand.rate}{The expansion rate of the bottom margin of the graph to incorporate footnote (numeric). The value of \code{1} indicates no expansion. If \code{NULL} (default), it is set to the number of lines in the footnote + 1.}

\item{show.plot}{Print the plot at the end of function (boulean). The default is \code{TRUE}.}

\item{flip.plot}{Flip the x and y axis of the plot. The default is \code{FALSE}. If \code{TRUE}, the variable names will be printed as columns and coefficients will be printed as rows. The original design does not intended for flipping the axis, thus flipping it may cause problems in the layout.}

\item{...}{Additional arguments passed to \code{\link[sandwich]{vcovHC}}, \code{\link[multiwayvcov]{cluster.vcov}}, or \code{\link[car]{Boot}} depending on the value of \code{vcov.est}.}
}
\value{
\code{ggplot} object without footnote (axis and theme settings can be added later). If footnote is added, then \code{gtable} object is the output. It is impossible to add \code{ggplot} elements to the \code{gtable} object. \code{gtable} plot can be viewed by using either \code{\link[grid]{grid.draw}()} or \code{\link[graphics]{plot}()} function.
}
\description{
Drawing the coefficient plots from either model result or coefficient table.
}
\examples{
## Load Data
library(pscl)
data(vote92)

## Recode Variables
vote92$voteBush <- as.numeric(
  factor(vote92$vote,levels=c("Clinton","Bush")))*1 - 1
vote92$bushdis <- sqrt(vote92$bushdis)
vote92$clintondis <- sqrt(vote92$clintondis)

## Estimate Logistic Regression
fm <- formula(voteBush ~ dem + rep +
                clintondis + bushdis +
                persfinance + natlecon)
m <- glm(fm, data = vote92,
         family = binomial("logit"))

## Basic Graph
plot_coef(m)

## Change X-axis to Odds Ratio
plot_coef(m, odds = TRUE)

## Export the Coefficients Table First, and import directly
cfci <- data.frame(cbind(coef(m), confint(m, level = 0.95)))
plot_coef(cfci)

## Estimate Model by Male and Female Subset
m_male <- glm(fm, data = vote92[vote92$female==0,],
         family = binomial("logit"))
m_female <- glm(fm, data = vote92[vote92$female==1,],
              family = binomial("logit"))

## Overlap Subsetted Results
plot_coef(list(m_male, m_female),
          m.names = c("Male", "Female"))

## Add Title and Custom Variable Names
vn <- c("(Intercept)",
        "Democrat","Republican",
        "Ideological Distance from Clinton",
        "Ideological Distance from Bush",
        "Retrospective Personal Finance",
        "Retrospective National Economy")
plot_coef(list(m_male, m_female),
          m.names = c("Male", "Female"),
          title = "Vote for Bush (1992)",
          custom.variable.names = vn)

## Overlap the Third Model & Facet Variables by Category & Add GOF Footnote
fn <- c("Constant",rep("Preference",4),rep("Evaluation",2))
fn <- factor(fn,levels=unique(fn))
plot_coef(list(m_male, m_female, m),
          m.names = c("Male", "Female", "All"),
          category.names = fn,
          title = "Vote for Bush (1992)",
          custom.variable.names = vn,
          footnote.gof = TRUE)

## Further Adding Facets 
plot_coef(list(m_male, m_female, m),
          m.names = c("Male", "Female", "ALL"),
          overlap.linetype.index = c(1,2,1),
          category.names = fn,
          facet.names = c("Gender Subsets","Gender Subsets","All"),
          title = "Vote for Bush (1992)",
          custom.variable.names = vn,
          footnote.gof = TRUE)


}
\references{
\url{http://www.surefoss.org/dataanalysis/plotting-odds-ratios-aka-a-forrestplot-with-ggplot2/} is where my initial idea come from.
}
