% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plotting Coefficients}
\usage{
plot_coef(x, direct = FALSE, order.variable = "original", odds = FALSE,
  drop.intercept = FALSE, drop.intercept.names = "(Intercept)",
  drop.variable.names = NULL, point.shape = 16, point.size = 1.5,
  confint.linetype = 1, confint.linewidth = 0.5, confint.height = 0.2,
  overlap.x.list = NULL, overlap.class.names = NULL,
  overlap.gapwidth = 0.5, overlap.shape.index = NULL,
  overlap.linetype.index = NULL, overlap.legend.position = "top",
  facet.category.names = NULL, facet.names.location = "left",
  facet.names.angle = 0, title = NULL, custom.variable.names = NULL,
  custom.themes = NULL)
}
\arguments{
\item{x}{The estimated model result object.}

\item{direct}{Import the result directly from the table (boulean). If \code{FALSE} (default), coefficients imported from estimated model object (\code{\link[stats]{coef}} and \code{\link[stats]{confint}} must be applicable). If \code{TRUE}, coefficients imported directly from coefficients table (rows=variables, columns=(coefficient, lower CI, upper CI), row names = variable names).}

\item{order.variable}{Order of coefficients in the plot. \code{"original"} (default) preserves the original order of the variable. \code{"coeforder"} plots by the discending order of the coefficient size. \code{"asis"} use the default \code{ggplot} setting.}

\item{odds}{Use odds ratio instead of coefficient in the output (boulean). The default is \code{FALSE}. If \code{TRUE}, the exponent of the coefficients will be plotted.}

\item{drop.intercept}{Drop the intercept from the plot (boulean). If \code{FALSE} (default), intercept included in the plot.}

\item{drop.intercept.names}{The name(s) of intercept (character/character vector). Needed if \code{drop.intercept} is \code{TRUE}. This is used to identify and eliminate intercept variables from the output. Default value is \code{"(Intercept)"}.}

\item{drop.variable.names}{The name(s) of additional variables to drop (character/character vector) from the ouput. The default is \code{NULL}.}

\item{point.shape}{Shape of the point outputs (numeric/character). The default is \code{16} (filled circle).}

\item{point.size}{Size of point outputs (numeric). The default is \code{1.5}.}

\item{confint.linetype}{The line type of confidence interval outputs (numeric). The default is \code{1}.}

\item{confint.linewidth}{The line width of confidence interval outputs (numeric). The default is \code{0.5}.}

\item{confint.height}{The height of the vertical line added to the edge of confidence interval outputs. The default is \code{0.2}.}

\item{overlap.x.list}{The list of additional estimated model result to be overlapped with the first result (list). Each element of the list must have the same set of variables and same class as the \code{x}. The default is \code{NULL}.}

\item{overlap.class.names}{The set of names that identifies overlapping \code{x}s (character vector). Needed if \code{overlap.x.list} is not \code{NULL}. The length of the vector must correspond with the length of \code{overlap.x.list} + 1. The first element of the vector is the name for \code{x}, then from the second element, the name order must correspond with the order in \code{overlap.x.list}.}

\item{overlap.gapwidth}{The gap between overlapped ouputs (number). The default value is \code{0.5}.}

\item{overlap.shape.index}{The index of shapes for overlapped point ouputs. Must be in the same length as \code{overlap.x.list} + 1. The first element of the vector is the shape for \code{x}, then from the second element, the order must correspond with the order in \code{overlap.x.list}. If \code{NULL}, \code{point.shape} is applied to all classes.}

\item{overlap.linetype.index}{The index of line types for overlapped confidence interval ouputs. Must be in the same length as \code{overlap.x.list} + 1. The first element of the vector is the shape for \code{x}, then from the second element, the order must correspond with the order in \code{overlap.x.list}. If \code{NULL}, the number corresponding with the order is assigned to each class.}

\item{overlap.legend.position}{The position of the legend for overlapping classess. See \code{legend.position} in ggplot theme for possible values. The default is \code{"top"}.}

\item{facet.category.names}{The categories of variables (factor). If not \code{NULL}, the output provides the separate panels for variables in each category. The length of the vector must much with the number of variables in \code{x} (This is considered BEFORE the application of \code{drop.intercept} and \code{drop.variable.names}. Just insert \code{NA} for those variables to be dropped.).}

\item{facet.names.location}{The location of facetted category names. Either \code{"left"} or \code{"right"}. The default is \code{"left"}.}

\item{facet.names.angle}{The angle of facetted category names (numeric). The default is \code{0} (horizontal).}

\item{title}{Plot title (character). The default is to include no title.}

\item{custom.variable.names}{List of alternative variable names in the output (character vector). The default is \code{NULL}. This is applied AFTER \code{drop.intercept} and \code{drop.variable.names} are applied, thus you don't need the names for dropped variables.}

\item{custom.themes}{ggplot themes that overrides the default theme. The default is \code{NULL}.}
}
\value{
\code{ggplot} object (axis and theme settings can be added later).
}
\description{
Drawing the coefficient plots from either model result or coefficient table. Borrowed the idea of odds ratio plotting from the code in \url{http://www.surefoss.org/dataanalysis/plotting-odds-ratios-aka-a-forrestplot-with-ggplot2/}.
}
