% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plotting Coefficients}
\usage{
plot_coef(m, m.names = NULL, order.variable = "original", odds = FALSE,
  vcov.est = NULL, sims.est = NULL, show.ci = TRUE, level.ci = 0.95,
  drop.intercept = FALSE, drop.intercept.names = "(Intercept)",
  drop.variable.names = NULL, point.shape = 16, point.size = 1.5,
  ci.linetype = 1, ci.size = 0.5, ci.height = 0.2,
  overlap.gapwidth = 0.5, overlap.shape.index = NULL,
  overlap.linetype.index = NULL, overlap.legend.position = "bottom",
  facet.category.names = NULL, facet.names.location = "left",
  facet.names.angle = 0, title = NULL, ytitle = NULL,
  custom.variable.names = NULL, custom.x.title = NULL,
  custom.x.lim = NULL, custom.x.breaks = NULL, custom.themes = NULL,
  footnote.gof = FALSE, footnote.gof.nobs = "right",
  footnote.gof.extracts = c("AIC", "Adj. R$^2$", "Log Likelihood"),
  footnote.gof.reorder = NULL, footnote.gof.linebreak = TRUE,
  custom.footnote = NULL, footnote.fontsize = NULL,
  footnote.fontcol = "black", footnote.align = "right",
  footnote.distance.from.side = 0.05, footnote.distance.from.bottom = 0.75,
  footnote.bottom.expand.rate = NULL, show.plot = TRUE, flip.plot = FALSE)
}
\arguments{
\item{m}{Single or the list of model object or the \code{matrix} or \code{data.frame} object. If it's a list all elements of the list should be the same class. If model object, \code{\link[stats]{coef}} and \code{\link[lmtest]{coefci}} must be applicable to the object. If  \code{matrix} or \code{data.frame}, then it should include following three columns and row names. Each row should correspond to each variable.
\itemize{
 \item{Row Names: }{Name of Variables}
 \item{First Column: }{Coefficient}
 \item{Second Column: }{Lower limit of confidence interval}
 \item{Third Column: }{Upper limit of confidence interval}
}}

\item{m.names}{The set of names that identifies each element in \code{m}. Considered if \code{m} is a list of models. The length of the vector must correspond with the length of \code{m}. If \code{NULL} (default), each element \code{i} is temporarily named as \code{Model i}.}

\item{order.variable}{Order of coefficients in the plot(character/numeric vector). \code{"original"} (default) preserves the original order of the variable. \code{"cfdescend"} plots by the descending order of coefficient values. \code{"cfascend"} plots by the ascending order of coefficient values. \code{"asis"} use the default \code{ggplot} setting.}

\item{odds}{Use odds ratio instead of coefficient in the output (boulean). The default is \code{FALSE}. If \code{TRUE}, the exponent of the coefficients will be plotted.}

\item{vcov.est}{Single or the list of the alternative variance-covariance matrix to be used for the model object \code{m}. Ignored if \code{NULL} (default) or the \code{m} is not model object. Must have the same length as \code{m} if it is a list.}

\item{sims.est}{Number of iterations if simulation method is used to calculate mean coefficient value and confidence interval. If \code{NULL} (default), the analytical method is used to calculate confidence intervals.}

\item{show.ci}{Show confidence interval (boulean). The default is \code{TRUE}.}

\item{level.ci}{The level used for confidence interval (numeric: 0-1). The default is \code{0.95}.}

\item{drop.intercept}{Drop the intercept from the plot (boulean). If \code{FALSE} (default), intercept included in the plot.}

\item{drop.intercept.names}{The name(s) of intercept (character/character vector). Needed if \code{drop.intercept} is \code{TRUE}. This is used to identify and eliminate intercept variables from the output. Default value is \code{"(Intercept)"}.}

\item{drop.variable.names}{The name(s) of additional variables to drop (character/character vector) from the ouput. The default is \code{NULL}.}

\item{point.shape}{Shape of the point outputs (numeric/character). The default is \code{16} (filled circle).}

\item{point.size}{Size of point outputs (numeric). The default is \code{1.5}.}

\item{ci.linetype}{The line type of confidence interval outputs (numeric). The default is \code{1}.}

\item{ci.size}{The line width of confidence interval outputs (numeric). The default is \code{0.5}.}

\item{ci.height}{The height of the vertical line added to the edge of confidence interval outputs. The default is \code{0.2}.}

\item{overlap.gapwidth}{The gap between overlapped ouputs (number). The default value is \code{0.5}.}

\item{overlap.shape.index}{The index of shapes for overlapped point ouputs. Must be in the same length as \code{overlap.m.list} + 1. The first element of the vector is the shape for \code{m}, then from the second element, the order must correspond with the order in \code{overlap.m.list}. If \code{NULL}, \code{point.shape} is applied to all classes.}

\item{overlap.linetype.index}{The index of line types for overlapped confidence interval ouputs. Must be in the same length as \code{overlap.m.list} + 1. The first element of the vector is the shape for \code{m}, then from the second element, the order must correspond with the order in \code{overlap.m.list}. If \code{NULL}, the number corresponding with the order is assigned to each class.}

\item{overlap.legend.position}{The position of the legend for overlapping classess. See \code{legend.position} in ggplot theme for possible values. The default is \code{"bottom"}.}

\item{facet.category.names}{The categories of variables (factor). If not \code{NULL}, the output provides the separate panels for variables in each category. The length of the vector must much with the number of variables in \code{m} (This is considered BEFORE the application of \code{drop.intercept} and \code{drop.variable.names}. Just insert \code{NA} for those variables to be dropped.).}

\item{facet.names.location}{The location of facetted category names. Either \code{"left"} or \code{"right"}. The default is \code{"left"}.}

\item{facet.names.angle}{The angle of facetted category names (numeric). The default is \code{0} (horizontal).}

\item{title}{Plot title (character). The default is to include no title.}

\item{ytitle}{Y axis title (character). The default is to include no axis title.}

\item{custom.variable.names}{List of alternative variable names in the output (character vector). The default is \code{NULL}. This is applied AFTER \code{drop.intercept} and \code{drop.variable.names} are applied, thus you don't need the names for dropped variables.}

\item{custom.x.title}{Custom name for the X axis title (character). The default is \code{NULL}.}

\item{custom.x.lim}{Custom limit for the X axis. The default is \code{NULL}.}

\item{custom.x.breaks}{Custom breaks for the X axis. The default is \code{NULL}.}

\item{custom.themes}{ggplot themes that overrides the default theme. The default is \code{NULL}.}

\item{footnote.gof}{Include GOF measures in the footnote (boulean). The default is \code{FALSE}. If \code{TRUE}, footnote with GOF measures are added to the plot by \code{\link{extract_gofchr}} and \code{\link{plot_footnote}} function, and the function exports \code{gtable} object. Note that \code{gtable} object is less customizable than \code{ggplot} object.}

\item{footnote.gof.nobs}{The position of "number of observation (N)" to be printed. The default is \code{"right"} (the end of the sentence). \code{"left"} (the start of the sentence) and \code{"none"} (do not print number of observation) are also available.}

\item{footnote.gof.extracts}{GOF measures to be inluded if \code{footnote.gof == TRUE}. See \code{\link{extract_gofchr}} documentation for more details.}

\item{footnote.gof.reorder}{The numeric vector specifying the alternative order of GOF  (e.g., if there are three GOFs, and you want to flip the order of first and second GOF, then input \code{c(2,1,3)}). The length of the vector must correspond with the number of exported GOFs (it may not be the same as the length of \code{footnote.gof.extracts}). If \code{NULL} (default), the original order is used. The order does not include the number of observation (if exported).}

\item{footnote.gof.linebreak}{Include linebreak between models (boulean). If \code{TRUE} (default), linebreak will be included between models. If \code{FALSE}, the ouput will be single line. Only applicable when there are more than one model.}

\item{custom.footnote}{Custom footnote (character). The default is \code{NULL}. If assigned, footnote are added to the plot by \code{\link{plot_footnote}} function, and the function exports \code{gtable} object. Note that \code{gtable} object is less customizable than \code{ggplot} object. If it is also the case that \code{footnote.gof == TRUE}, custom footnote will be added as the new line after the GOF footnote.}

\item{footnote.fontsize}{The size of font. If \code{NULL} (default), the size is set to the the font size in \code{text} setting of ggplot theme - 1.}

\item{footnote.fontcol}{The color of the font. The default is \code{"black"}.}

\item{footnote.align}{The alignment of the footnote text. Either \code{"right"} or \code{"left"}.}

\item{footnote.distance.from.side}{The horizontal distance of notes from the edge of graph space by the proportion of graph width (numeric: 0-1). The default is \code{0.05}. The distance is measured from the side specified in \code{align}.}

\item{footnote.distance.from.bottom}{The vertical distance of notes from the bottom of graph bottom by the proportion of bottom graph margin height (numeric: 0-1). The default is \code{0.75}.}

\item{footnote.bottom.expand.rate}{The expansion rate of the bottom margin of the graph to incorporate footnote (numeric). The value of \code{1} indicates no expansion. If \code{NULL} (default), it is set to the number of lines in the footnote + 1.}

\item{show.plot}{Print the plot at the end of function (boulean). The default is \code{TRUE}.}

\item{flip.plot}{Flip the x and y axis of the plot. The default is \code{FALSE}. If \code{TRUE}, the variable names will be printed as columns and coefficients will be printed as rows. The original design does not intended for flipping the axis, thus flipping it may cause problems in the layout.}
}
\value{
\code{ggplot} object without footnote (axis and theme settings can be added later). If footnote is added, then \code{gtable} object is the output. It is impossible to add \code{ggplot} elements to the \code{gtable} object. \code{gtable} plot can be viewed by using either \code{\link[grid]{grid.draw}()} or \code{\link[graphics]{plot}()} function.
}
\description{
Drawing the coefficient plots from either model result or coefficient table.
}
\examples{
## Load Data
library(pscl)
data(vote92)

## Recode Variables
vote92$voteBush <- as.numeric(
  factor(vote92$vote,levels=c("Clinton","Bush")))*1 - 1
vote92$bushdis <- sqrt(vote92$bushdis)
vote92$clintondis <- sqrt(vote92$clintondis)

## Estimate Logistic Regression
fm <- formula(voteBush ~ dem + rep +
                clintondis + bushdis +
                persfinance + natlecon)
m <- glm(fm, data = vote92,
         family = binomial("logit"))

## Basic Graph
plot_coef(m)

## Change X-axis to Odds Ratio
plot_coef(m, odds = TRUE)

## Export the Coefficients Table First, and import directly
cfci <- data.frame(cbind(coef(m), confint(m, level = 0.95)))
plot_coef(cfci)

## Estimate Model by Male and Female Subset
m_male <- glm(fm, data = vote92[vote92$female==0,],
         family = binomial("logit"))
m_female <- glm(fm, data = vote92[vote92$female==1,],
              family = binomial("logit"))

## Overlap Subsetted Results
plot_coef(list(m_male, m_female),
          m.names = c("Male", "Female"))

## Add Title and Custom Variable Names
vn <- c("(Intercept)",
        "Democrat","Republican",
        "Ideological Distance from Clinton",
        "Ideological Distance from Bush",
        "Retrospective Personal Finance",
        "Retrospective National Economy")
plot_coef(list(m_male, m_female),
          m.names = c("Male", "Female"),
          title = "Vote for Bush (1992)",
          custom.variable.names = vn)

## Overlap the Third Model & Facet Variables by Category & Add GOF Footnote
fn <- c("Constant",rep("Preference",4),rep("Evaluation",2))
fn <- factor(fn,levels=unique(fn))
plot_coef(list(m_male, m_female, m),
          m.names = c("Male", "Female", "All"),
          facet.category.names = fn,
          title = "Vote for Bush (1992)",
          custom.variable.names = vn,
          footnote.gof = TRUE)

}
\references{
\url{http://www.surefoss.org/dataanalysis/plotting-odds-ratios-aka-a-forrestplot-with-ggplot2/} is where my initial idea come from.
}
