% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_interact.R
\name{plot_interact}
\alias{plot_interact}
\title{Plot Interaction Effects}
\usage{
plot_interact(
  interactprof,
  name.var2 = NULL,
  label.var2 = "name.var2",
  label.y = "auto",
  hline0 = TRUE,
  addprof = NULL,
  name.est = "Mean",
  show.ci = TRUE,
  name.lowerCI = "lowerCI",
  name.upperCI = "upperCI",
  type.est = "auto",
  type.ci = "auto",
  name.facet.x = NULL,
  name.facet.y = NULL,
  name.color = NULL,
  name.linetype = NULL,
  name.fill = NULL,
  name.shape = NULL,
  line.width = 0.7,
  barpoint.gapwidth = 0.5,
  point.size = 2,
  errorbar.width = 0.5,
  ribbon.alpha = 0.5,
  label.color = "name.color",
  label.linetype = "name.linetype",
  label.fill = "name.fill",
  label.shape = "name.shape",
  titletxt = NULL
)
}
\arguments{
\item{interactprof}{Conditional coefficients in either \code{simuinteract} (exported from \code{\link{simu_interact}} 
function), \code{data.frame} or \code{matrix} class. If in \code{data.frame} or \code{matrix},
it should have rows as predcion cases and columns for mean/median prediction, 
lower and upper confidence intervals (if \code{show.ci==TRUE}), and values for 
predictors that varies across profiles.}

\item{name.var2}{Column name for the main variable that conditions the coeffient of target variable (plotted on x axis). 
If not given, the information will be extracted from \code{interactprof} (only when it is \code{simuinteract} object).}

\item{label.var2}{X axis label. If \code{"name.var2"} (default), use the variable name defined in \code{name.var2}.}

\item{label.y}{Y axis label. If \code{"auto"} (default) and \code{interactprof} being 
\code{simuinteract} object, the label is automatically determined.}

\item{hline0}{If \code{TRUE}, add horizontal line at the value of zero.}

\item{addprof}{Optional \code{data.frame} that adds or replaces values of conditions. 
Must have the same number of cases as \code{interactprof}. If 
variable name is found in \code{interactprof}, values are replaced. If not, additional variable 
is added to \code{interactprof}. Added variables can be used in \code{name.var2}, 
\code{name.color}, \code{name.linetype}, \code{name.fill}, and \code{name.shape}.
Define variable as \code{factor} if you want to control order of appearance in legend.}

\item{name.est}{Column name for mean/median conditional coefficient estiamte.}

\item{show.ci}{If \code{TRUE} (default), plot confidence intervals.}

\item{name.lowerCI}{Column name for the upper limit of confidence interval (applied if \code{show.ci==TRUE}).}

\item{name.upperCI}{Column name for the upper limit of confidence interval (applied if \code{show.ci==TRUE}).}

\item{type.est}{Plotting type of estimates. If \code{"auto"}, the type is determined automatically chosen by 
the class of the variable defined by \code{name.var2}. If not, choose manually from \code{"line"}, \code{"bar"}, or \code{"point"}.}

\item{type.ci}{Plotting type of confidence intervals. If \code{"auto"}, the type is determined automatically chosen by 
the class of the variable defined by \code{name.var2}. If not, choose manually from \code{"errorbar"} (Error Bars) or \code{"ribbon"} (Appropriate if \code{type.est == "line"}).}

\item{name.facet.x}{Column name for the predictor that controls facets horizontally (optional).}

\item{name.facet.y}{Column name for the predictor that controls facets vertically (optional).}

\item{name.color}{Column name for the predictor that controls line colors (optional).}

\item{name.linetype}{COlumn name for the predictor that controls line types (optional).}

\item{name.fill}{Column name for the predictor that controls fill colors (optional).}

\item{name.shape}{Column name for the predictor that controls point shapes (optional).}

\item{line.width}{The width of line in line plot.}

\item{barpoint.gapwidth}{The gap between bars/points if plot is clustered.}

\item{point.size}{The size of points in point plot.}

\item{errorbar.width}{The width of errorbar.}

\item{ribbon.alpha}{The transparency of ribbon plot.}

\item{label.color}{Optional label if \code{name.color==TRUE}. If \code{"name.color"}, the value assigned to \code{name.color} will be used for name in legend.}

\item{label.linetype}{Optional label if \code{name.linetype==TRUE}. If \code{"name.linetype"}, the value assigned to \code{name.linetype} will be used for name in legend.}

\item{label.fill}{Optional label if \code{name.fill==TRUE}. If \code{"name.fill"}, the value assigned to \code{name.fill} will be used for name in legend.}

\item{label.shape}{Optional label if \code{name.shape==TRUE}. If \code{"name.shape"}, the value assigned to \code{name.shape} will be used for name in legend.}

\item{titletxt}{Title Text.}
}
\value{
\code{ggplot} object (axis and theme settings can be added later).
}
\description{
Drawing interaction plot. This function is a wrapper for \code{\link{plot_simu}} function.
}
\examples{

## Load Data
library(pscl)
data(vote92)

## Recode Variables
vote92$voteBush <- as.numeric(
factor(vote92$vote,levels=c("Clinton","Bush")))*1 - 1
vote92$bushdis <- sqrt(vote92$bushdis)
vote92$clintondis <- sqrt(vote92$clintondis)

## Estimate Logistic Regression with Interaction
fm <- formula(voteBush ~ dem*clintondis + 
                rep + bushdis +
                persfinance + natlecon)
m <- glm(fm, data = vote92,
         family = binomial("logit"))
         
# Moving Values
moveprof <- data.frame(clintondis = seq(0,4,length=50))
# Simulation
interactprof <- simu_interact(m, "dem", moveprof=moveprof,
                              var1.label = "Being Democrat",
                              y.label = "Bush Vote")
# Plot
plot_interact(interactprof, 
              label.var2="Ideological Distance from Clinton")

# Alternative Way
interactprof <- simu_interact(m, "dem", var2 = "clintondis",
                              var1.label = "Being Democrat",
                              y.label = "Bush Vote")
plot_interact(interactprof,
              label.var2="Ideological Distance from Clinton")

}
